+++ R/ggplot_box_fun.R
@@ -135,7 +135,7 @@ theme_box <-
 #' @param wrap_x_title Number of characters to wrap the x title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_y_title Number of characters to wrap the y title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shinyapp, isMobile should be specified as input$isMobile.
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -185,7 +185,14 @@ ggplot_box <- function(data,
                        wrap_x_title = 50,
                        wrap_y_title = 50,
                        wrap_caption = 80,
-                       isMobile = FALSE){
+                       isMobile = NULL){
+  
+  if(is.null(isMobile)){
+    shiny <- shiny::isRunning()
+    if(shiny == FALSE) isMobile <- FALSE
+    else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+    else isMobile <- FALSE
+  }
   
   data <- dplyr::ungroup(data)
   x_var <- rlang::enquo(x_var) #categorical var
@@ -302,7 +309,7 @@ ggplot_box <- function(data,
         title = stringr::str_wrap(title, 40),
         subtitle = stringr::str_wrap(subtitle, 40),
         x = stringr::str_wrap(x_title, 20),
-        y = stringr::str_wrap(y_title, 20),
+        y = stringr::str_wrap(y_title, 30),
         caption = stringr::str_wrap(caption, 50)
       ) +
       coord_flip() +
@@ -342,7 +349,7 @@ ggplot_box <- function(data,
 #' @param wrap_x_title Number of characters to wrap the x title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_y_title Number of characters to wrap the y title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shinyapp, isMobile should be specified as input$isMobile.
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -386,7 +393,14 @@ ggplot_box_facet <-
            wrap_x_title = 50,
            wrap_y_title = 50,
            wrap_caption = 80,
-           isMobile = FALSE){
+           isMobile = NULL){
+    
+    if(is.null(isMobile)){
+      shiny <- shiny::isRunning()
+      if(shiny == FALSE) isMobile <- FALSE
+      else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+      else isMobile <- FALSE
+    }
     
     data <- dplyr::ungroup(data)
     x_var <- rlang::enquo(x_var) #categorical var
@@ -521,7 +535,7 @@ ggplot_box_facet <-
           title = stringr::str_wrap(title, 40),
           subtitle = stringr::str_wrap(subtitle, 40),
           x = stringr::str_wrap(x_title, 20),
-          y = stringr::str_wrap(y_title, 20),
+          y = stringr::str_wrap(y_title, 30),
           caption = stringr::str_wrap(caption, 50)
         ) +
         facet_wrap(vars(!!facet_var), scales = facet_scales, ncol = 1) +