+++ R/ggplot_hbar_fun.R
@@ -138,7 +138,7 @@ theme_hbar <-
 #' @param wrap_x_title Number of characters to wrap the x title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_y_title Number of characters to wrap the y title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shinyapp, isMobile should be specified as input$isMobile.
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -186,7 +186,14 @@ ggplot_hbar <- function(data,
                         wrap_x_title = 50,
                         wrap_y_title = 50,
                         wrap_caption = 80,
-                        isMobile = FALSE){
+                        isMobile = NULL){
+  
+  if(is.null(isMobile)){
+    shiny <- shiny::isRunning()
+    if(shiny == FALSE) isMobile <- FALSE
+    else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+    else isMobile <- FALSE
+  }
   
   data <- dplyr::ungroup(data)
   x_var <- rlang::enquo(x_var) #numeric var
@@ -432,7 +439,7 @@ ggplot_hbar <- function(data,
 #' @param wrap_y_title Number of characters to wrap the y title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_col_title Number of characters to wrap the colour title to. Defaults to 25. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shinyapp, isMobile should be specified as input$isMobile.
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -492,7 +499,14 @@ ggplot_hbar_col <-
            wrap_y_title = 50,
            wrap_col_title = 25,
            wrap_caption = 80,
-           isMobile = FALSE){
+           isMobile = NULL){
+    
+    if(is.null(isMobile)){
+      shiny <- shiny::isRunning()
+      if(shiny == FALSE) isMobile <- FALSE
+      else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+      else isMobile <- FALSE
+    }
     
     data <- dplyr::ungroup(data)
     x_var <- rlang::enquo(x_var) #numeric var
@@ -748,7 +762,7 @@ ggplot_hbar_col <-
 #' @param wrap_x_title Number of characters to wrap the x title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_y_title Number of characters to wrap the y title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shinyapp, isMobile should be specified as input$isMobile.
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -801,7 +815,14 @@ ggplot_hbar_facet <-
            wrap_x_title = 50,
            wrap_y_title = 50,
            wrap_caption = 80,
-           isMobile = FALSE){
+           isMobile = NULL){
+    
+    if(is.null(isMobile)){
+      shiny <- shiny::isRunning()
+      if(shiny == FALSE) isMobile <- FALSE
+      else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+      else isMobile <- FALSE
+    }
     
     data <- dplyr::ungroup(data)
     y_var <- rlang::enquo(y_var) #categorical var
@@ -1074,7 +1095,7 @@ ggplot_hbar_facet <-
 #' @param wrap_y_title Number of characters to wrap the y title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_col_title Number of characters to wrap the colour title to. Defaults to 25. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shinyapp, isMobile should be specified as input$isMobile.
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -1133,7 +1154,14 @@ ggplot_hbar_col_facet <-
            wrap_y_title = 50,
            wrap_col_title = 25,
            wrap_caption = 80,
-           isMobile = FALSE){
+           isMobile = NULL){
+    
+    if(is.null(isMobile)){
+      shiny <- shiny::isRunning()
+      if(shiny == FALSE) isMobile <- FALSE
+      else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+      else isMobile <- FALSE
+    }
     
     data <- dplyr::ungroup(data)
     y_var <- rlang::enquo(y_var) #categorical var