+++ R/ggplot_line_fun.R
@@ -138,7 +138,7 @@ theme_line <-
 #' @param wrap_x_title Number of characters to wrap the x title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_y_title Number of characters to wrap the y title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shinyapp, isMobile should be specified as input$isMobile.
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -184,7 +184,14 @@ ggplot_line <- function(data,
                         wrap_x_title = 50,
                         wrap_y_title = 50,
                         wrap_caption = 80,
-                        isMobile = FALSE) {
+                        isMobile = NULL) {
+  
+  if(is.null(isMobile)){
+    shiny <- shiny::isRunning()
+    if(shiny == FALSE) isMobile <- FALSE
+    else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+    else isMobile <- FALSE
+  }
   
   data <- dplyr::ungroup(data)
   x_var <- rlang::enquo(x_var) #numeric var
@@ -356,7 +363,7 @@ ggplot_line <- function(data,
         title = stringr::str_wrap(title, 40),
         subtitle = stringr::str_wrap(subtitle, 40),
         x = stringr::str_wrap(x_title, 20),
-        y = stringr::str_wrap(y_title, 20),
+        y = stringr::str_wrap(y_title, 30),
         caption = stringr::str_wrap(caption, 50)
       )
   }
@@ -401,7 +408,7 @@ ggplot_line <- function(data,
 #' @param wrap_y_title Number of characters to wrap the y title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_col_title Number of characters to wrap the colour title to. Defaults to 25. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shinyapp, isMobile should be specified as input$isMobile.
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -453,7 +460,14 @@ ggplot_line_col <-
            wrap_y_title = 50,
            wrap_col_title = 25,
            wrap_caption = 80,
-           isMobile = FALSE) {
+           isMobile = NULL) {
+    
+    if(is.null(isMobile)){
+      shiny <- shiny::isRunning()
+      if(shiny == FALSE) isMobile <- FALSE
+      else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+      else isMobile <- FALSE
+    }
     
     data <- dplyr::ungroup(data)
     x_var <- rlang::enquo(x_var) #numeric var
@@ -650,7 +664,7 @@ ggplot_line_col <-
           title = stringr::str_wrap(title, 40),
           subtitle = stringr::str_wrap(subtitle, 40),
           x = stringr::str_wrap(x_title, 20),
-          y = stringr::str_wrap(y_title, 20),
+          y = stringr::str_wrap(y_title, 30),
           caption = stringr::str_wrap(caption, 50)
         )  +
         guides(col = guide_legend(ncol = 1, byrow = TRUE, title = stringr::str_wrap(col_title, 15)))
@@ -692,7 +706,7 @@ ggplot_line_col <-
 #' @param wrap_x_title Number of characters to wrap the x title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_y_title Number of characters to wrap the y title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shinyapp, isMobile should be specified as input$isMobile.
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -740,7 +754,14 @@ ggplot_line_facet <-
            wrap_x_title = 50,
            wrap_y_title = 50,
            wrap_caption = 80,
-           isMobile = FALSE) {
+           isMobile = NULL) {
+    
+    if(is.null(isMobile)){
+      shiny <- shiny::isRunning()
+      if(shiny == FALSE) isMobile <- FALSE
+      else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+      else isMobile <- FALSE
+    }
     
     data <- dplyr::ungroup(data)
     x_var <- rlang::enquo(x_var) #numeric var
@@ -931,7 +952,7 @@ ggplot_line_facet <-
           title = stringr::str_wrap(title, 40),
           subtitle = stringr::str_wrap(subtitle, 40),
           x = stringr::str_wrap(x_title, 20),
-          y = stringr::str_wrap(y_title, 20),
+          y = stringr::str_wrap(y_title, 30),
           caption = stringr::str_wrap(caption, 50)
         )  +
         facet_wrap(vars(!!facet_var), scales = facet_scales, ncol = 1)
@@ -981,7 +1002,7 @@ ggplot_line_facet <-
 #' @param wrap_y_title Number of characters to wrap the y title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_col_title Number of characters to wrap the colour title to. Defaults to 25. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shinyapp, isMobile should be specified as input$isMobile.
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -1037,7 +1058,14 @@ ggplot_line_col_facet <-
            wrap_y_title = 50,
            wrap_col_title = 25,
            wrap_caption = 80,
-           isMobile = FALSE) {
+           isMobile = NULL) {
+    
+    if(is.null(isMobile)){
+      shiny <- shiny::isRunning()
+      if(shiny == FALSE) isMobile <- FALSE
+      else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+      else isMobile <- FALSE
+    }
     
     data <- dplyr::ungroup(data)
     x_var <- rlang::enquo(x_var) #numeric var
@@ -1262,7 +1290,7 @@ ggplot_line_col_facet <-
           title = stringr::str_wrap(title, 40),
           subtitle = stringr::str_wrap(subtitle, 40),
           x = stringr::str_wrap(x_title, 20),
-          y = stringr::str_wrap(y_title, 20),
+          y = stringr::str_wrap(y_title, 30),
           caption = stringr::str_wrap(caption, 50)
         )  +
         facet_wrap(vars(!!facet_var), scales = facet_scales, ncol = 1) +