+++ R/ggplot_scatter_fun.R
@@ -139,7 +139,7 @@ theme_scatter <-
 #' @param wrap_x_title Number of characters to wrap the x title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_y_title Number of characters to wrap the y title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shinyapp, isMobile should be specified as input$isMobile.
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -184,7 +184,14 @@ ggplot_scatter <- function(data,
                            wrap_x_title = 50,
                            wrap_y_title = 50,
                            wrap_caption = 80,
-                           isMobile = FALSE) {
+                           isMobile = NULL) {
+  
+  if(is.null(isMobile)){
+    shiny <- shiny::isRunning()
+    if(shiny == FALSE) isMobile <- FALSE
+    else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+    else isMobile <- FALSE
+  }
   
   data <- dplyr::ungroup(data)
   x_var <- rlang::enquo(x_var) #numeric var
@@ -347,7 +354,7 @@ ggplot_scatter <- function(data,
         title = stringr::str_wrap(title, 40),
         subtitle = stringr::str_wrap(subtitle, 40),
         x = stringr::str_wrap(x_title, 20),
-        y = stringr::str_wrap(y_title, 20),
+        y = stringr::str_wrap(y_title, 30),
         caption = stringr::str_wrap(caption, 50)
       )
   }
@@ -398,7 +405,7 @@ ggplot_scatter <- function(data,
 #' @param wrap_y_title Number of characters to wrap the y title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_col_title Number of characters to wrap the colour title to. Defaults to 25. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shinyapp, isMobile should be specified as input$isMobile.
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -453,7 +460,14 @@ ggplot_scatter_col <-
            wrap_y_title = 50,
            wrap_col_title = 25,
            wrap_caption = 80,
-           isMobile = FALSE) {
+           isMobile = NULL) {
+    
+    if(is.null(isMobile)){
+      shiny <- shiny::isRunning()
+      if(shiny == FALSE) isMobile <- FALSE
+      else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+      else isMobile <- FALSE
+    }
     
     data <- dplyr::ungroup(data)
     x_var <- rlang::enquo(x_var) #numeric var
@@ -670,7 +684,7 @@ ggplot_scatter_col <-
           title = stringr::str_wrap(title, 40),
           subtitle = stringr::str_wrap(subtitle, 40),
           x = stringr::str_wrap(x_title, 20),
-          y = stringr::str_wrap(y_title, 20),
+          y = stringr::str_wrap(y_title, 30),
           caption = stringr::str_wrap(caption, 50)
         )  +
         guides(col = guide_legend(ncol = 1, byrow = TRUE, title = stringr::str_wrap(col_title, 15)))
@@ -713,7 +727,7 @@ ggplot_scatter_col <-
 #' @param wrap_x_title Number of characters to wrap the x title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_y_title Number of characters to wrap the y title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shinyapp, isMobile should be specified as input$isMobile.
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -759,7 +773,14 @@ ggplot_scatter_facet <-
            wrap_x_title = 50,
            wrap_y_title = 50,
            wrap_caption = 80,
-           isMobile = FALSE) {
+           isMobile = NULL) {
+    
+    if(is.null(isMobile)){
+      shiny <- shiny::isRunning()
+      if(shiny == FALSE) isMobile <- FALSE
+      else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+      else isMobile <- FALSE
+    }
     
     data <- dplyr::ungroup(data)
     x_var <- rlang::enquo(x_var) #numeric var
@@ -951,7 +972,7 @@ ggplot_scatter_facet <-
           title = stringr::str_wrap(title, 40),
           subtitle = stringr::str_wrap(subtitle, 40),
           x = stringr::str_wrap(x_title, 20),
-          y = stringr::str_wrap(y_title, 20),
+          y = stringr::str_wrap(y_title, 30),
           caption = stringr::str_wrap(caption, 50)
         )  +
         facet_wrap(vars(!!facet_var), scales = facet_scales, ncol = 1)
@@ -1007,7 +1028,7 @@ ggplot_scatter_facet <-
 #' @param wrap_y_title Number of characters to wrap the y title to. Defaults to 50. Not applicable where isMobile equals TRUE.
 #' @param wrap_col_title Number of characters to wrap the colour title to. Defaults to 25. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shinyapp, isMobile should be specified as input$isMobile.
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -1069,7 +1090,14 @@ ggplot_scatter_col_facet <-
            wrap_y_title = 50,
            wrap_col_title = 25,
            wrap_caption = 80,
-           isMobile = FALSE) {
+           isMobile = NULL) {
+    
+    if(is.null(isMobile)){
+      shiny <- shiny::isRunning()
+      if(shiny == FALSE) isMobile <- FALSE
+      else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+      else isMobile <- FALSE
+    }
     
     data <- dplyr::ungroup(data)
     x_var <- rlang::enquo(x_var) #numeric var
@@ -1330,7 +1358,7 @@ ggplot_scatter_col_facet <-
           title = stringr::str_wrap(title, 40),
           subtitle = stringr::str_wrap(subtitle, 40),
           x = stringr::str_wrap(x_title, 20),
-          y = stringr::str_wrap(y_title, 20),
+          y = stringr::str_wrap(y_title, 30),
           caption = stringr::str_wrap(caption, 50)
         )  +
         guides(col = guide_legend(ncol = 1, byrow = TRUE, title = stringr::str_wrap(col_title, 15))) +