+++ R/ggplot_sf_fun.R
@@ -104,7 +104,7 @@ theme_sf <-
 #' @param wrap_title Number of characters to wrap the title to. Defaults to 70. Not applicable where isMobile equals TRUE.
 #' @param wrap_subtitle Number of characters to wrap the subtitle to. Defaults to 80. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shiny app, use input$isMobile if your app is able to retreive this input. A method to do this is described at https://g3rv4.com/2017/08/shiny-detect-mobile-browsers
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -126,7 +126,14 @@ ggplot_sf <- function(data,
                       wrap_title = 70,
                       wrap_subtitle = 80,
                       wrap_caption = 80,
-                      isMobile = FALSE) {
+                      isMobile = NULL) {
+  
+  if(is.null(isMobile)){
+    shiny <- shiny::isRunning()
+    if(shiny == FALSE) isMobile <- FALSE
+    else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+    else isMobile <- FALSE
+  }
   
   if (class(data)[1] != "sf") stop("Please use an sf object as data input")
   if (is.na(sf::st_crs(data))) stop("Please assign a coordinate reference system")
@@ -226,7 +233,7 @@ ggplot_sf <- function(data,
 #' @param wrap_subtitle Number of characters to wrap the subtitle to. Defaults to 80. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
 #' @param wrap_col_title Number of characters to wrap the colour title to. Defaults to 25. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shiny app, use input$isMobile if your app is able to retreive this input. A method to do this is described at https://g3rv4.com/2017/08/shiny-detect-mobile-browsers
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -274,7 +281,14 @@ ggplot_sf_col <- function(data,
                           wrap_subtitle = 80,
                           wrap_col_title = 25,
                           wrap_caption = 80,
-                          isMobile = FALSE) {
+                          isMobile = NULL) {
+  
+  if(is.null(isMobile)){
+    shiny <- shiny::isRunning()
+    if(shiny == FALSE) isMobile <- FALSE
+    else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+    else isMobile <- FALSE
+  }
   
   col_var <- rlang::enquo(col_var)
   
@@ -462,7 +476,7 @@ ggplot_sf_col <- function(data,
 #' @param wrap_title Number of characters to wrap the title to. Defaults to 70. Not applicable where isMobile equals TRUE.
 #' @param wrap_subtitle Number of characters to wrap the subtitle to. Defaults to 80. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shiny app, use input$isMobile if your app is able to retreive this input. A method to do this is described at https://g3rv4.com/2017/08/shiny-detect-mobile-browsers
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -489,7 +503,14 @@ ggplot_sf_facet <- function(data,
                             wrap_title = 70,
                             wrap_subtitle = 80,
                             wrap_caption = 80,
-                            isMobile = FALSE) {
+                            isMobile = NULL) {
+  
+  if(is.null(isMobile)){
+    shiny <- shiny::isRunning()
+    if(shiny == FALSE) isMobile <- FALSE
+    else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+    else isMobile <- FALSE
+  }
   
   facet_var <- rlang::enquo(facet_var) #categorical var
   
@@ -622,7 +643,7 @@ ggplot_sf_facet <- function(data,
 #' @param wrap_subtitle Number of characters to wrap the subtitle to. Defaults to 80. Not applicable where isMobile equals TRUE.
 #' @param wrap_col_title Number of characters to wrap the colour title to. Defaults to 25. Not applicable where isMobile equals TRUE.
 #' @param wrap_caption Number of characters to wrap the caption to. Defaults to 80. Not applicable where isMobile equals TRUE.
-#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to FALSE. In a shiny app, use input$isMobile if your app is able to retreive this input. A method to do this is described at https://g3rv4.com/2017/08/shiny-detect-mobile-browsers
+#' @param isMobile Whether the plot is to be displayed on a mobile device. Defaults to NULL, which is FALSE unless run inside an app with the mobileDetect function available.
 #' @return A ggplot object.
 #' @export
 #' @examples
@@ -666,7 +687,14 @@ ggplot_sf_col_facet <- function(data,
                                 wrap_subtitle = 80,
                                 wrap_col_title = 25,
                                 wrap_caption = 80,
-                                isMobile = FALSE) {
+                                isMobile = NULL) {
+  
+  if(is.null(isMobile)){
+    shiny <- shiny::isRunning()
+    if(shiny == FALSE) isMobile <- FALSE
+    else if(shiny == TRUE & exists("mobileDetect")) isMobile <- input$isMobile
+    else isMobile <- FALSE
+  }
   
   col_var <- rlang::enquo(col_var)
   facet_var <- rlang::enquo(facet_var) #categorical var